+++ Application/LightSwitch/advanced/backend/controllers/FinancialController.php
@@ -51,9 +51,14 @@ class FinancialController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+         if( Yii::$app->user->can('Core-Members'))
+        {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+           throw new ForbiddenHttpException('Are you an authorized Member?');
+        }
     }
 
     /**
@@ -75,7 +80,7 @@ class FinancialController extends Controller
                 ]);
             }
         } else {
-           throw new ForbiddenHttpException('Are you a Core Member?');
+           throw new ForbiddenHttpException('Are you a authorized Member?');
         }
     }
 
@@ -87,24 +92,29 @@ class FinancialController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+         if( Yii::$app->user->can('Core-Members'))
+        {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
 
-            // get the instance of the uploaded file
-             $imageName = $model->file_name;
-             $model->file = UploadedFile::getInstance($model,'file');
-             $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
+                // get the instance of the uploaded file
+                 $imageName = $model->file_name;
+                 $model->file = UploadedFile::getInstance($model,'file');
+                 $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension);
 
-             // save the file path in the db column
-             $model->files = 'uploads/'.$imageName.'.'.$model->file->extension;
-             $model->save();
+                 // save the file path in the db column
+                 $model->files = 'uploads/'.$imageName.'.'.$model->file->extension;
+                 $model->save();
 
-            return $this->redirect(['view', 'id' => $model->id]);
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+           throw new ForbiddenHttpException('Are you a authorized Member?');
         }
     }
 
@@ -116,9 +126,14 @@ class FinancialController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if( Yii::$app->user->can('Core-Members'))
+        {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+           throw new ForbiddenHttpException('Are you a authorized Member?');
+        }
     }
 
     /**
