+++ Application/LightSwitch/advanced/backend/controllers/ProjectController.php
@@ -9,6 +9,7 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\web\UploadedFile;
+use yii\web\ForbiddenHttpException;
 /**
  * ProjectController implements the CRUD actions for Project model.
  */
@@ -48,9 +49,14 @@ class ProjectController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if( Yii::$app->user->can('Core-Members'))
+        {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        } else {
+           throw new ForbiddenHttpException('Are you a authorized Member?');
+        }
     }
 
     /**
@@ -60,14 +66,19 @@ class ProjectController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Project();
+        if( Yii::$app->user->can('Core-Members'))
+        {
+            $model = new Project();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('create', [
-                'model' => $model,
-            ]);
+           throw new ForbiddenHttpException('Are you a authorized Member?');
         }
     }
 
@@ -79,26 +90,31 @@ class ProjectController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if( Yii::$app->user->can('Core-Members'))
+        {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
 
 
-             // get the instance of the uploaded file
-             $imageName = $model->file_name;
-             $model->file = UploadedFile::getInstance($model,'file');
-             $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension );
+                 // get the instance of the uploaded file
+                 $imageName = $model->file_name;
+                 $model->file = UploadedFile::getInstance($model,'file');
+                 $model->file->saveAs('uploads/'.$imageName.'.'.$model->file->extension );
 
-             // save the file path in the db column
-             $model->files = 'uploads/'.$imageName.'.'.$model->file->extension;
-             $model->save();
+                 // save the file path in the db column
+                 $model->files = 'uploads/'.$imageName.'.'.$model->file->extension;
+                 $model->save();
 
-             
-            return $this->redirect(['view', 'id' => $model->id]);
+                 
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                'model' => $model,
-            ]);
+           throw new ForbiddenHttpException('Are you a authorized Member?');
         }
     }
 
@@ -110,9 +126,14 @@ class ProjectController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if( Yii::$app->user->can('Core-Members'))
+        {   
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+           throw new ForbiddenHttpException('Are you a authorized Member?');
+        }
     }
 
     /**
